+++ packages/custom_lint/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased patch
+
+- Sort lints by severity in the command line (thanks to @kuhnroyal)
+- Fix watch mode not quitting with `q` (thanks to @kuhnroyal)
+
 ## 0.5.3 - 2023-08-29
 
 - The command line now supports ignoring warnings/infos with `--no-fatal-warnings`/`--no-fatal-infos` (thanks to @yamarkz)